# chore(logging): minor perf improvements and cleanup