--- a/mod/log/pkg/phuslu/buffer.go
+++ b/mod/log/pkg/phuslu/buffer.go
@@ -27,12 +27,6 @@ type byteBuffer struct {
 	Bytes []byte
 }
 
-// Write writes to the byte buffer.
-func (b *byteBuffer) Write(bytes []byte) (int, error) {
-	b.Bytes = append(b.Bytes, bytes...)
-	return len(bytes), nil
-}
-
 // byteBufferPool is a pool of byte buffers.
 //
 //nolint:gochecknoglobals // buffer pool
@@ -42,7 +36,14 @@ var byteBufferPool = sync.Pool{
 	},
 }
 
-func resetBuffer(b *byteBuffer) {
+// Write writes to the byte buffer.
+func (b *byteBuffer) Write(bytes []byte) (int, error) {
+	b.Bytes = append(b.Bytes, bytes...)
+	return len(bytes), nil
+}
+
+// Reset resets the byte buffer.
+func (b *byteBuffer) Reset() {
 	if b.Bytes != nil {
 		b.Bytes = b.Bytes[:0]
 	} else {