--- a/mod/log/pkg/phuslu/formatter.go
+++ b/mod/log/pkg/phuslu/formatter.go
@@ -96,9 +96,10 @@ func (f *Formatter) Format(
 	f.ensureLineBreak(buffer)
 
 	if args.Stack != "" {
-		buffer.Bytes = append(buffer.Bytes, args.Stack...)
 		if args.Stack[len(args.Stack)-1] != '\n' {
-			buffer.Bytes = append(buffer.Bytes, '\n')
+			buffer.Bytes = append(buffer.Bytes, append([]byte(args.Stack), '\n')...)
+		} else {
+			buffer.Bytes = append(buffer.Bytes, args.Stack...)
 		}
 	}