Before:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkPhusluLoggerInfo$ github.com/berachain/beacon-kit/testing/benchmarks

goos: darwin
goarch: arm64
pkg: github.com/berachain/beacon-kit/testing/benchmarks
=== RUN   BenchmarkPhusluLoggerInfo
BenchmarkPhusluLoggerInfo
BenchmarkPhusluLoggerInfo-16             1414178               854.2 ns/op          930 B/op         16 allocs/op
PASS
ok      github.com/berachain/beacon-kit/testing/benchmarks    3.196s
```

After:
```
go test -benchmem -run=^$ -bench ^BenchmarkPhusluLoggerInfo$ github.com/berachain/beacon-kit/testing/benchmarks
goos: darwin
goarch: arm64
pkg: github.com/berachain/beacon-kit/testing/benchmarks
BenchmarkPhusluLoggerInfo-16    	 2808595	       441.2 ns/op	     662 B/op	       5 allocs/op
PASS
ok  	github.com/berachain/beacon-kit/testing/benchmarks	2.046s
```



<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Refactor**
  - Enhanced `Formatter` type with new methods for initialization and log message formatting, improving structure and control flow.
  
- **Style**
  - Capitalized the first letter of log messages across various modules for consistency and clarity.

- **New Features**
  - Introduced `NewFormatter` method for creating a `Formatter` with default settings.
  - Added fields `QuoteString` and `EndWithMessage` to `Formatter` to customize log message formatting.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->