@ocnc 2024-06-19
> gimme that 200ns

oh fuck for context i wrote this without looking at your benchmark LMAO